Avoid shallow clone
Chartpress detects the latest commit that changed a directory
or file when determining the version and tag to use for charts
and images. This means that shallow clones should not be used
because if the last commit that changed a relevant file is
outside the shallow commit range, the wrong chart version
and image tag will be assigned.