Automatically resize the filesystem if necessary
Currently, if you grow the underlying disk to expand storage,
you also have to manually run `xfs_growfs` to expand the filesystem.
This adds heavy operational load here, as you need to then use
both terraform (or the cloud console) *and* the kubectl commandline.

With this, we simply run `xfs_growfs` in a loop every 5 minutes.
So at most 5min after the underlying disk has changed, that space
is available for users.